refactored ticket and customer class as well as the testing for these classes

+++ models/film.rb
@@ -7,7 +7,7 @@ class Film
   def initialize(options)
     @id = options['id'].to_i
     @title = options['title']
-    @price = options['price'].to_i
+    @price = options['price'].to_f
   end 
 
   def save
